requirements.txt
streamlit
pandasai
python-dotenv
bs4
numpy
streamlit-modal
